MAIN/requirements.txt
Django==4.1.7
psycopg2-binary
django-debug-toolbar
djangorestframework
redis
django-rq
pillow
django-cors-headers
pygments